requirements.txt
streamlit
pandas
scikit-learn
python-pptx